Support custom version or coordinates for Compose compiler (#207)
Since this project does not build a Kotlin compiler plugin, we should not be forced to do a release for new versions of Kotlin. This property will allow consumers to upgrade their Kotlin version by specifying a newer JetBrains Compose compiler version, or a totally different set of Maven coordinates for a custom Compose compiler artifact.