Add AnnotatedString with SpanStyle for text customization (#227)
* Add AnnotatedString with SpanStyle for text customization

This is a simplified version of AnnotatedString and SpanStyle from Jetpack Compose. There is no ParagraphStyle support. SpanStyle contains fewer properties.

The tests are also taken from Jetpack Compose, but simplified.

The API has been saved for using AnnotatedString and creating and styling it, for quick understanding if you know Jetpack Compose.

The code has two Compose Text functions, two DrawScope.DrawText methods, two TextLayout inheritors, in which the difference is only in the accepted value - either String or AnnotatedString. This is done for optimization, but in theory it may be unnecessary.

Update the jest example using AnnotatedString.

* Add MosaicComposable to Text function with AnnotatedString parameter

* Change TextLayout access modifier to internal