Using print for output
I tried to compare the option before and the option after, I didn't really see the difference, maybe there are cases when it is noticeable? I'll be glad if you tell me.

I wanted to make an issue with the question first, but decided that it would be faster, because there are minimal changes and immediately, if everything is good, we can apply them.