Launch different phases separately depending on need
A simple implementation that remembers the readings of the states when calling the layout, drawing blocks separately, and then with the help of a global observer checks which states have changed, if those that were read during layout or drawing have changed and composition is not going on now, then we run the layout or drawing phase separately.

Since composition does not occur in case of some changes, therefore, when completed in the `MosaicComposition#awaitComplete` method, we additionally expect one frame to be rendered, if necessary.